moved DependencyNode class from 'internal.core.deployment.parser' to 'api.config' package
