+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/core/deployment/AppCreator.java
@@ -25,6 +25,7 @@ import com.google.common.collect.SetMultimap;
 import org.wso2.carbon.uuf.api.Placeholder;
 import org.wso2.carbon.uuf.api.config.ComponentManifest;
 import org.wso2.carbon.uuf.api.config.Configuration;
+import org.wso2.carbon.uuf.api.config.DependencyNode;
 import org.wso2.carbon.uuf.api.reference.AppReference;
 import org.wso2.carbon.uuf.api.reference.ComponentReference;
 import org.wso2.carbon.uuf.api.reference.FileReference;
@@ -46,7 +47,6 @@ import org.wso2.carbon.uuf.exception.UUFException;
 import org.wso2.carbon.uuf.internal.core.auth.SessionRegistry;
 import org.wso2.carbon.uuf.internal.core.deployment.parser.ComponentManifestParser;
 import org.wso2.carbon.uuf.internal.core.deployment.parser.ConfigurationParser;
-import org.wso2.carbon.uuf.internal.core.deployment.parser.DependencyNode;
 import org.wso2.carbon.uuf.internal.core.deployment.parser.DependencyTreeParser;
 import org.wso2.carbon.uuf.internal.util.NameUtils;
 import org.wso2.carbon.uuf.spi.RenderableCreator;
